+++ muxers/mplex/CHANGELOG.md
@@ -1,6 +1,7 @@
 # unreleased
 
-- Add `set_protocol_name` to MplexConfig
+- Add `fn set_protocol_name(&mut self, protocol_name: &'static [u8])` to MplexConfig
+- Add `fn protocol_name(&self)` to MplexConfig
 
 # 0.30.0 [2021-11-01]